+++ events/messageDelete.js
@@ -7,15 +7,22 @@ module.exports = {
         let messagedelete = new Discord.EmbedBuilder()
           .setTitle(`Message deleted`)
           .addFields(
-            { name: `Message Content`, value: `${messageDelete.content}`},
+            { name: `Message Content`, value: `${messageDelete.content || "*No text content*"}`},
             {name: `Message by`, value: `${messageDelete.author}`},
-            {name: `Link:`, value: `${messageDelete.url}`}
+            {name: `Link:`, value: `${messageDelete.url || "N/A"}`}
           )
           .setColor("Orange");
+
+        if (messageDelete.attachments.size > 0) {
+            const attachmentUrls = messageDelete.attachments.map(att => att.url).join("\n")
+            messagedelete.addFields({ name: "Attachments", value: attachmentUrls })
+        }
+
     if (!logchannel) {
-      console.warn(`Log channel with ID ${LOG_CHANNEL} not found in guild ${member.guild.id}`);
+      console.warn(`Log channel with ID ${LOGS_CHANNEL} not found in guild ${messageDelete.guild.id}`);
       return; 
     }
         await logchannel.send({embeds: [messagedelete] });
   }
 };
+